AgentWithMemoryStorageReasoning/requirements.txt
agno
python-dotenv 
google-genai
duckduckgo-search #ddgs(in future)
lancedb 
tantivy 
sqlalchemy
pypdf
pylance
pandas
yfinance
newspaper4k 
lxml_html_clean
arxiv
googlesearch-python
pycountry

